move role to user types files
