+++ src/types/general.ts
@@ -17,11 +17,6 @@ export enum ItemStatus {
   WontDo = "WontDo",
 }
 
-export enum Role {
-  User = "User",
-  Admin = "Admin"
-}
-
 export function stringToActionStatus(statusString: string): ItemStatus {
   const status = statusString.trim();
